fix(windows): handle keyboard package not downloaded

+++ windows/src/desktop/kmshell/main/Keyman.System.UpdateStateMachine.pas
@@ -767,6 +767,7 @@ begin
     // IdleState to wait 'CheckPeriod' before trying again
     TKeymanSentryClient.Client.MessageEvent(Sentry.Client.SENTRY_LEVEL_ERROR,
     'Error Updates not downloaded after 3 attempts');
+    bucStateContext.RemoveCachedFiles;
     ChangeState(IdleState);
   end
   else
@@ -996,17 +997,28 @@ function InstallingState.DoInstallPackage(PackageFileName: String): Boolean;
 var
   FPackage: IKeymanPackageFile2;
 begin
-  Result := True;
-  FPackage := kmcom.Packages.GetPackageFromFile(PackageFileName)
-    as IKeymanPackageFile2;
-  FPackage.Install2(True);
-  // Force overwrites existing package and leaves most settings for it intact
+  Result := False;
+  try
+    FPackage := kmcom.Packages.GetPackageFromFile(PackageFileName)
+      as IKeymanPackageFile2;
+      // Force overwrites existing package and leaves most settings for it intact
+    FPackage.Install2(True);
+    Result := True;
+  except
+    on E:Exception do
+    begin
+      TKeymanSentryClient.ReportHandledException(E,
+        'Failed to install keyboard package');
+    end;
+  end;
+
   FPackage := nil;
-
-  kmcom.Refresh;
-  kmcom.Apply;
-
-  System.SysUtils.DeleteFile(PackageFileName);
+  if Result then
+  begin
+    kmcom.Refresh;
+    kmcom.Apply;
+    System.SysUtils.DeleteFile(PackageFileName);
+  end;
 
 end;
 
@@ -1023,10 +1035,9 @@ begin
     PackageFullPath := SavePath + Params.Packages[i].FileName;
     if not FileExists(PackageFullPath) then
     begin
-      TKeymanSentryClient.Client.MessageEvent(Sentry.Client.SENTRY_LEVEL_ERROR,
-      'File does not exist:"' + PackageFullPath + '"');
       Continue;
     end;
+
     if not DoInstallPackage(PackageFullPath) then // I2742
     begin
       TKeymanSentryClient.Client.MessageEvent(Sentry.Client.SENTRY_LEVEL_ERROR,
